Password Validation moved to own file
